Connected the local database to account
I have create a local database and uploaded the database file along with new login h/cpp/UI files, and a new database connection class (databasereaderwriter). I also edited Account class and made sure it's not a singleton anymore, as well as maindashboard stuff. More edits to come, but this is the very strong basis.
IMPORTANT: In order to make sure you can connect to database and run it appropriately, make sure to edit the databasereaderwriter.cpp file, where YOU WILL NEED TO EDIT the constructor that sets the path to the database, so make sure you specify the path before you attempt to login. There's 3 users, but use Username: Alex and Password: helloworld.