Implementation of 'Edit Transaction'
NOTE:
UI should be working perfectly fine, the way I went about it was the program would prompt the user to enter a row # to edit, then would pass that value to an "editTransaction()" function which would proceed to update the vector index.

I was unable to figure out a way to properly format this, as to whether or not I would just do "transactions[index] = editValue" or if I should delete the index and insert into the position.

Other than that, the UI implementation should be running fine.